requirements.txt
django
django-rest-framework
django-countries
names
drf-yasg